requirements.txt
PyYaml
selenium
